backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
SQLAlchemy==2.0.22
pydantic==1.10.11
pandas==2.1.1
python-dateutil==2.8.2